Merge master w real posts (#32)
* add: preparing redis

added the redis gem. I am trying to merge master and postsRealtime.

* Posts realtime merging with a branch that cloned master (#31)

* add: upgraded rails 4 into the latest patch and fixed depreciation warnings from devise.

* fix: upgrading from Rails 4 to 5

The upgrade was needed because it will allow developers to use websockets
for real-time updates.

* add: there is real-time post updates

I still need to incorporate partials and use ajax to update the
partials paramers.

* fix: Correctly renders a partial

When a user posts a new post users will get that data appended
to the view via a partial. No long-polling used.

* fix: fixed the order of new data

Fixed the order of how jQuery was appending the partials
to the view.

* add: setting up redis to go

adding setup in order for redis to work in
heroku.

* fix: removed files from redisToGo setup

removed files that were being used for setting up
redis with postgre.

* fix: forgot one merge conflict.

* fix: forgot to remove head

* fix: fixed migration error

added [5.1] to the migration file.

* fix: removed puts used in the controller

* add: added redis support for yml

* fix: removed sqlLite from the GemFile

caused an error when deploying to heroku

* fix: removed merge conflict syntax

there was merge conflict syntax that was placed after the "end"
of the class.

* fix: returned sqlite4 into the development environment

* fix: fixed codeclimate warnings

* fix: reran bundle install

update the gems installed